requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database
pymongo==4.6.3
motor==3.5.1

# Data Validation
pydantic==2.7.4
pydantic-settings==2.3.4

# AI/LLM
groq==0.4.1

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# File Processing
python-multipart==0.0.9
pdfplumber==0.11.0
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.1
tenacity==8.5.0
httpx==0.27.0
python-slugify==8.0.4
